+++ src/core/include/cesium/omniverse/UsdUtil.h
@@ -19,6 +19,8 @@ class CesiumPolygonRasterOverlay;
 class CesiumSession;
 class CesiumTileset;
 class CesiumWebMapServiceRasterOverlay;
+class CesiumTileMapServiceRasterOverlay;
+class CesiumWebMapTileServiceRasterOverlay;
 class UsdGeomBasisCurves;
 class UsdGeomXformable;
 class UsdGeomXformOp;
@@ -49,6 +51,8 @@ namespace cesium::omniverse::UsdUtil {
 glm::dvec3 usdToGlmVector(const pxr::GfVec3d& vector);
 glm::fvec3 usdToGlmVector(const pxr::GfVec3f& vector);
 glm::dmat4 usdToGlmMatrix(const pxr::GfMatrix4d& matrix);
+glm::dquat usdToGlmQuat(const pxr::GfQuatd& quat);
+glm::fquat usdToGlmQuat(const pxr::GfQuatf& quat);
 std::array<glm::dvec3, 2> usdToGlmExtent(const pxr::GfRange3d& extent);
 
 pxr::GfVec3d glmToUsdVector(const glm::dvec3& vector);
@@ -122,6 +126,10 @@ pxr::CesiumPolygonRasterOverlay
 getCesiumPolygonRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 pxr::CesiumWebMapServiceRasterOverlay
 getCesiumWebMapServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
+pxr::CesiumTileMapServiceRasterOverlay
+getCesiumTileMapServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
+pxr::CesiumWebMapTileServiceRasterOverlay
+getCesiumWebMapTileServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 pxr::CesiumGeoreference getCesiumGeoreference(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 pxr::CesiumGlobeAnchorAPI getCesiumGlobeAnchor(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 pxr::CesiumIonServer getCesiumIonServer(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
@@ -137,6 +145,8 @@ bool isCesiumRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPat
 bool isCesiumIonRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 bool isCesiumPolygonRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 bool isCesiumWebMapServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
+bool isCesiumTileMapServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
+bool isCesiumWebMapTileServiceRasterOverlay(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 bool isCesiumGeoreference(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 bool isCesiumIonServer(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
 bool isCesiumCartographicPolygon(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path);
@@ -147,14 +157,16 @@ bool isUsdMaterial(const pxr::UsdStageWeakPtr& pStage, const pxr::SdfPath& path)
 
 glm::dvec3 getTranslate(const pxr::UsdGeomXformOp& translateOp);
 glm::dvec3 getRotate(const pxr::UsdGeomXformOp& rotateOp);
+glm::dquat getOrient(const pxr::UsdGeomXformOp& orientOp);
 glm::dvec3 getScale(const pxr::UsdGeomXformOp& scaleOp);
 void setTranslate(pxr::UsdGeomXformOp& translateOp, const glm::dvec3& translate);
 void setRotate(pxr::UsdGeomXformOp& rotateOp, const glm::dvec3& rotate);
+void setOrient(pxr::UsdGeomXformOp& orientOp, const glm::dquat& orient);
 void setScale(pxr::UsdGeomXformOp& scaleOp, const glm::dvec3& scale);
 
 struct TranslateRotateScaleOps {
     pxr::UsdGeomXformOp translateOp;
-    pxr::UsdGeomXformOp rotateOp;
+    pxr::UsdGeomXformOp rotateOrOrientOp;
     pxr::UsdGeomXformOp scaleOp;
     MathUtil::EulerAngleOrder eulerAngleOrder;
 };